fix: no pending data message
