+++ src/components/blocks/StagingDataTable.js
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux';
 import styled, { withTheme } from 'styled-components';
 import { TableCellHeaderText, TableCellText } from '../typography';
 import { convertPascalCaseToSentenceCase } from '../../utils/stringUtils';
-import { MinusIcon, H3 } from '..';
+import { MinusIcon } from '..';
 import { TableDrawer } from './';
 import { useWindowSize } from '../hooks/useWindowSize';
 
@@ -74,13 +74,6 @@ const Td = styled('td')`
 
 const StagingDataTableContainer = styled('div')`
   height: 100%;
-`;
-
-const NoDataMessageContainer = styled('div')`
-  display: flex;
-  height: 100%;
-  justify-content: center;
-  align-items: center;
 `;
 
 const ChangeGroupHeader = ({ headings, appStore }) => {
@@ -202,11 +195,6 @@ const StagingDataTable = withTheme(({ headings, data, deleteStagingData }) => {
   return (
     <StagingDataTableContainer ref={ref}>
       <div style={{ height: `${height}px`, overflow: 'auto' }}>
-        {data.length === 0 && (
-          <NoDataMessageContainer>
-            <H3>No staged data at this time</H3>
-          </NoDataMessageContainer>
-        )}
         {data &&
           headings &&
           data.map((changeGroup, index) => (
